requirements.txt
streamlit
google-generativeai
dotenv
pdfplumber
python-docx